+++ Test-MAME-YM2608/tone_file.cpp
@@ -1,6 +1,6 @@
 #include "tone_file.hpp"
 #include <fstream>
-#include "types.h"
+#include <cstdint>
 
 ToneFile::ToneFile() {}
 
@@ -18,27 +18,27 @@ std::unique_ptr<Tone> ToneFile::read(const std::string file)
         ifs.read(buffer, 4);
         str = std::string(buffer, 4);
         if (str == "TONE") {
-            uint32 hdOffset;
-            ifs.read(reinterpret_cast<char*>(&hdOffset), sizeof(uint32));
-            size_t strLen;
-            ifs.read(reinterpret_cast<char*>(&strLen), sizeof(uint32));
+			uint32_t hdOffset;
+			ifs.read(reinterpret_cast<char*>(&hdOffset), sizeof(uint32_t));
+			uint32_t strLen;
+			ifs.read(reinterpret_cast<char*>(&strLen), sizeof(uint32_t));
             ifs.read(buffer, strLen);
             tone->name = std::string(buffer, strLen);
 
             ifs.seekg(4 + sizeof(hdOffset) + hdOffset, std::ios_base::beg);
-            ifs.read(reinterpret_cast<char*>(&tone->AL), sizeof(uint8));
-            ifs.read(reinterpret_cast<char*>(&tone->FB), sizeof(uint8));
+			ifs.read(reinterpret_cast<char*>(&tone->AL), sizeof(uint8_t));
+			ifs.read(reinterpret_cast<char*>(&tone->FB), sizeof(uint8_t));
             for (auto& op : tone->op) {
-                ifs.read(reinterpret_cast<char*>(&op.AR), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.DR), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.SR), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.RR), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.SL), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.TL), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.KS), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.ML), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.DT), sizeof(uint8));
-                ifs.read(reinterpret_cast<char*>(&op.AM), sizeof(uint8));
+				ifs.read(reinterpret_cast<char*>(&op.AR), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.DR), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.SR), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.RR), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.SL), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.TL), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.KS), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.ML), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.DT), sizeof(uint8_t));
+				ifs.read(reinterpret_cast<char*>(&op.AM), sizeof(uint8_t));
             }
         }
     }
@@ -52,24 +52,24 @@ void ToneFile::write(const std::string file, Tone* tone)
 
     if (ofs) {
         ofs.write("TONE", 4);
-        uint32 hdOffset = sizeof(uint32) + tone->name.size();
-        ofs.write(reinterpret_cast<char*>(&hdOffset), sizeof(uint32));
-        uint32 nameLen = tone->name.size();
-        ofs.write(reinterpret_cast<char*>(&nameLen), sizeof(uint32));
+		uint32_t hdOffset = sizeof(uint32_t) + tone->name.size();
+		ofs.write(reinterpret_cast<char*>(&hdOffset), sizeof(uint32_t));
+		uint32_t nameLen = tone->name.size();
+		ofs.write(reinterpret_cast<char*>(&nameLen), sizeof(uint32_t));
         ofs.write(tone->name.c_str(), tone->name.size());
-        ofs.write(reinterpret_cast<char*>(&tone->AL), sizeof(uint8));
-        ofs.write(reinterpret_cast<char*>(&tone->FB), sizeof(uint8));
+		ofs.write(reinterpret_cast<char*>(&tone->AL), sizeof(uint8_t));
+		ofs.write(reinterpret_cast<char*>(&tone->FB), sizeof(uint8_t));
         for (auto& op : tone->op) {
-            ofs.write(reinterpret_cast<char*>(&op.AR), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.DR), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.SR), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.RR), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.SL), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.TL), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.KS), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.ML), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.DT), sizeof(uint8));
-            ofs.write(reinterpret_cast<char*>(&op.AM), sizeof(uint8));
+			ofs.write(reinterpret_cast<char*>(&op.AR), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.DR), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.SR), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.RR), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.SL), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.TL), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.KS), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.ML), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.DT), sizeof(uint8_t));
+			ofs.write(reinterpret_cast<char*>(&op.AM), sizeof(uint8_t));
         }
     }
 }
